add twitch llm max tokens vs. story tokens
allow setting twitch llm usage different than main stories.

reduce story length max tokens for llm story messages.

+++ src/openai_api.rs
@@ -21,7 +21,7 @@ pub struct Message {
 pub struct OpenAIRequest<'a> {
     pub model: &'a str,
     pub messages: Vec<Message>,
-    pub max_tokens: &'a i32,        // add this field to the request struct
+    pub max_tokens: &'a usize,      // add this field to the request struct
     pub temperature: &'a f32,       // add this field to the request struct
     pub top_p: &'a f32,             // add this field to the request struct
     pub presence_penalty: &'a f32,  // add this field to the request struct
